Counting sort
Counting sort is working in O(n). Conditions for work: all values are integers and non-negative, the maximum possible value is given. Program creates an array with quantity of element's repetitions at index, that corresponds to this element's value in the 1st array (created at the beginning). Final output is a consistent record of 2nd array's indexex, repeting n times ('n' is a value at current index).